refactor(supabase): split pipeline recovery mutations into scraper-free import-recovery.ts
The decouple in c288a68d was too blunt: it dropped the whole pipeline module
from the mutations barrel and `export * from ./sources` from the root barrel.
That removed exports the WEB legitimately uses at runtime (Node handles cheerio
fine; only Jest's ts-jest transform chokes on it) — breaking the web typecheck
(7 missing exports), web test shard 2, and the local dev server.

Root cause: pipeline.ts co-located scraper-dependent stage runners
(runSyncStage/runImportStage/runCompileStage, which import ../sources) with
pure-DB recovery/admin verbs (eventKeyFor, deleteSourceEvent, excludeSourceEvent,
clearExclusion, resetStuckEvents, requeueFailedEvents, forceImportEvent). The
recovery verbs need NONE of the scraper imports, but sharing a module meant
importing them dragged cheerio into the consumer's graph.

Fix: move the pure verbs to import-recovery.ts (imports only TypedClient) and
re-export them from the mutations barrel — web/Jest get them with zero cheerio.
The stage runners stay in pipeline.ts behind the /pipeline subpath (edge-fn only).
ALL_VALID_FORMATS (sources/limitless/format.ts — a pure, import-free Set) is now
re-exported explicitly from the root barrel instead of via `export * from sources`.

Verified: web limitless-display-status + actions/pipeline 62/62, package pipeline
21/21, web+supabase typecheck clean.
